# docs/swagger.yaml
basePath: /v1
definitions:
  dao.Account:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      password:
        type: string
      token:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  dto.AccountCreateReq:
    properties:
      paswd:
        maxLength: 255
        minLength: 8
        type: string
      uname:
        maxLength: 16
        type: string
    required:
    - paswd
    - uname
    type: object
  dto.AccountCreateResp:
    properties:
      id:
        type: integer
      username:
        type: string
    type: object
  dto.AccountLoginReq:
    properties:
      paswd:
        maxLength: 255
        minLength: 8
        type: string
      uname:
        maxLength: 16
        type: string
    required:
    - paswd
    - uname
    type: object
  dto.AccountLoginResp:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  dto.AccountResp:
    properties:
      age:
        type: integer
      fullname:
        type: string
      gender:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  dto.AccountUpdateReq:
    properties:
      email:
        type: string
      id:
        type: integer
      paswd:
        maxLength: 255
        minLength: 8
        type: string
      uname:
        maxLength: 16
        type: string
    required:
    - email
    - id
    - paswd
    - uname
    type: object
...
  dto.ErrorResponse:
    properties:
      errors: {}
      message:
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  dto.SuccessResponse-any:
    properties:
      data: {}
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.SuccessResponse-dto_AccountCreateResp:
    properties:
      data:
        $ref: '#/definitions/dto.AccountCreateResp'
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.SuccessResponse-dto_AccountLoginResp:
    properties:
      data:
        $ref: '#/definitions/dto.AccountLoginResp'
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.SuccessResponse-dto_AccountResp:
    properties:
      data:
        $ref: '#/definitions/dto.AccountResp'
      message:
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:3000
info:
  contact:
    email: mark.p.e.muhammad@gmail.com
    name: Mark Muhammad
  description: This is a boilerplate project, please update accordingly.
  license:
    name: MIT
  title: Base API Service
  version: "1.0"
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: Retrieves a list of accounts based on the provided filter
      operationId: get-accounts
      parameters:
      - in: query
        minimum: 1
        name: l
        type: integer
      - in: query
        name: q
        type: string
      - in: query
        minimum: 0
        name: s
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of accounts
          schema:
            items:
              $ref: '#/definitions/dao.Account'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a list of accounts
      tags:
      - default
  /accounts/:
    post:
      consumes:
      - application/json
      description: Create a new account using the provided credentials
      parameters:
      - description: Account creation request
        in: body
        name: cred
        required: true
        schema:
          $ref: '#/definitions/dto.AccountCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_AccountCreateResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new account
  /accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an account using the provided ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete an account
    get:
      consumes:
      - application/json
      description: Retrieves an account by its ID
      operationId: get-account-by-id
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Account found
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_AccountResp'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get account by ID
    put:
      consumes:
      - application/json
      description: Updates an account based on the provided ID and request body
      operationId: update-account
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Account data
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.AccountUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Updated account
          schema:
            $ref: '#/definitions/dao.Account'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an account
      tags:
      - default
  /accounts/login:
    post:
      consumes:
      - application/json
      description: Account login using username & password combination.
      parameters:
      - description: Credential
        in: body
        name: cred
        required: true
        schema:
          $ref: '#/definitions/dto.AccountLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_AccountLoginResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Account login
  /accounts/profile:
    get:
      consumes:
      - application/json
      description: Get the profile information of the logged-in account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse-dto_AccountResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get logged-in account profile
...
securityDefinitions:
  BearerAuth:
    description: Bearer auth containing JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"